# server/swagger.yaml
openapi: 3.0.0
info:
  title: Project Management System
  description: MongoDB ExpressJS NodeJS REST API
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/
    description: Localhost development server 
components:
  securitySchemes:
    ApiKeyAuth:        # arbitrary name for the security scheme
      type: apiKey
      in: header       # can be "header", "query" or "cookie"
      name: auth-token
  
  schemas:
    project:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
    
    User:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string

paths:

  /user/register:
    post:
      summary: Register a new User
      description: Register a new User in the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"

      responses:
        201:
          description: User created succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  _id:
                    type: string

  /user/login:
    post:
      summary: Login a User
      description: Login a User in order to get JWT token (auth-token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      
      responses:
        200:
          description: User logged in succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  data:
                    type: string


  /project:
    get:
      summary: Retrieves a list of projects
      description: Retrieves a list of projects as JSON objects.
      responses:
        200:
          description: A list of projects JSON objects in an array.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/project"
    
    post:
      summary: Create a new project
      description: Create a new project
      security:
        - ApiKeyAuth: []      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/project"

      responses:
        201:
          description: project created succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/project"

  /project/{id}:
    get:
      summary: Specific project
      description: Retrieves a specific project based on it id.
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string
      responses:
        200:
          description: A project in the format of a JSON object.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/project"
   
    put:
      summary: Updates a specific project
      description: Updates a specific project based on it id
      security:
        - ApiKeyAuth: []            
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/project"

      responses:
        201:
          description: project updated succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/project"

    delete:
      summary: Deletes a specific project
      description: Deletes a specific project based on it id
      security:
        - ApiKeyAuth: []            
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string      

      responses:
        201:
          description: project deleted succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/project"